platform/ai-gateway/requirements.txt
# CustomERP AI Gateway Dependencies
# Python 3.11+

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# AI Providers
google-generativeai==0.8.0
openai>=1.40.0

# Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.9.0

# HTTP Client
httpx==0.27.0

# JSON Schema Validation
jsonschema==4.23.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.0
